+++ src/evaluate/mod.rs
@@ -29,10 +29,16 @@ pub enum ScopeKind {
     Component, // Object
 }
 
+#[derive(Debug, Clone, PartialEq)]
+pub(crate) struct VariableBinding {
+    pub value: Value,
+    pub is_const: bool,
+}
+
 /// Internal structure used for scope handling
 #[derive(PartialEq)]
 pub(crate) struct Scope {
-    pub variables: FxHashMap<String, Value>,
+    pub variables: FxHashMap<String, VariableBinding>,
     pub kind: ScopeKind,
 }
 
@@ -89,7 +95,7 @@ impl Evaluator {
                 TopLevelItem::Import(imp) => self.handle_import(imp.clone(), &mut root_nodes)?,
                 TopLevelItem::ComponentDef(def) => self.registry.register_component(def.clone()),
                 TopLevelItem::FnDef(def) => self.registry.register_function(def.clone()),
-                TopLevelItem::Stmt(Stmt::Global(name, expr)) => {
+                TopLevelItem::Stmt(Stmt::Const(name, expr)) => {
                     // We evaluate globals now so they are available in Loop 2
                     let val = self.eval_expr(expr)?;
                     self.registry.globals.insert(name.clone(), val);